Software/requirements.txt
asgiref==3.3.0
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
django-pwa==1.0.10
six==1.15.0
gunicorn==20.0.4
whitenoise==5.2.0
python-docx==0.8.10
django-htmlmin==0.11.0
django-cors-headers==3.7.0